# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "منصة خبرات - Khibrat HR SaaS API",
        "description": "التوثيق الرسمي لكافة واجهات برمجة التطبيقات (APIs) لمنصة خبرات لإدارة الموارد البشرية",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "سيرفر التطوير المحلي"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "المصادقة (Authentication)"
                ],
                "summary": "تسجيل دخول المستخدم للمنصة",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "hr@khibrat.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تسجيل الدخول بنجاح وإعادة الـ Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "type": "object"
                                                },
                                                "company": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|lhA7G..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "بيانات الدخول غير صحيحة"
                    },
                    "403": {
                        "description": "الحساب غير نشط"
                    }
                }
            }
        },
        "/api/auth/forgot-password": {
            "post": {
                "tags": [
                    "المصادقة (Authentication)"
                ],
                "summary": "طلب رابط إعادة تعيين كلمة المرور (نسيان كلمة المرور)",
                "operationId": "cff9b52a898f73f39d684747bd69c4e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "hr@khibrat.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم توليد رمز إعادة التعيين بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset link has been generated."
                                        },
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "hr@khibrat.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "البريد الإلكتروني غير مسجل في النظام"
                    }
                }
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "tags": [
                    "المصادقة (Authentication)"
                ],
                "summary": "إعادة تعيين كلمة المرور الجديدة باستخدام الـ Token",
                "operationId": "6af92e090f994419a2295111b391a205",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "hr@khibrat.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abcdef123456..."
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new_password_123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث كلمة المرور بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password updated successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "hr@khibrat.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "الـ Token غير صحيح أو منتهي الصلاحية"
                    }
                }
            }
        },
        "/api/auth/complete-first-login": {
            "post": {
                "tags": [
                    "المصادقة (Authentication)"
                ],
                "summary": "تغيير كلمة المرور الإلزامية عند تسجيل الدخول الأول للمنصة",
                "operationId": "b954f9b16adb04877cbe5d6a41900a9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "New@Pass2026"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "New@Pass2026"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث كلمة المرور وإغلاق حالة الدخول الأول بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Initial password changed successfully. Welcome to Khibrat!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "المستخدم غير مصادق عليه (Unauthenticated)"
                    },
                    "422": {
                        "description": "خطأ في التحقق من المدخلات (مثال: عدم تطابق كلمة المرور أو ضعفها)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "أدخلي الـ Token الذي حصلتِ عليه من الـ Login هنا لتفعيل الصلاحيات",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}